Move PaginatedItemOverview to common
