+++ graylog2-web-interface/src/components/common/PaginatedItemOverview/PaginatedItem.jsx
