+++ graylog2-web-interface/src/components/common/PaginatedItemOverview/index.jsx
@@ -1,0 +1,20 @@
+// @flow strict
+import PaginatedItemOverview, {
+  defaultPageInfo,
+  type PaginationInfo,
+  type PaginatedListType,
+  type DescriptiveItem,
+} from './PaginatedItemOverview';
+
+export type {
+  PaginationInfo,
+  PaginatedListType,
+  DescriptiveItem,
+};
+
+export {
+  defaultPageInfo,
+  PaginatedItemOverview,
+};
+
+export default PaginatedItemOverview;
