+++ graylog2-web-interface/src/components/users/UserDetails/RolesSection.jsx
@@ -5,8 +5,11 @@ import { useState } from 'react';
 import UserNotification from 'util/UserNotification';
 import User from 'logic/users/User';
 import { AuthzRolesActions } from 'stores/roles/AuthzRolesStore';
+import PaginatedItemOverview, {
+  type PaginationInfo,
+  type PaginatedListType,
+} from 'components/common/PaginatedItemOverview';
 
-import PaginatedItemOverview, { type PaginationInfo, type PaginatedListType } from '../PaginatedItemOverview';
 import SectionComponent from '../SectionComponent';
 
 type Props = {
